requirements.txt
altair==6.0.0
anyio==4.12.1
attrs==25.4.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deep-translator==1.11.4
exceptiongroup==1.3.1
ffmpeg-python==0.2.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
indic_transliteration==2.3.76
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.16.0
networkx==3.4.2
nltk==3.9.2
numba==0.63.1
numpy==2.2.6
openai-whisper==20250625
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.5
pyarrow==23.0.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
roman==5.2
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.3
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.1.0
typer==0.21.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
watchdog==6.0.0
wordcloud==1.9.6
